+++ src/cmd/plot.rs
@@ -26,7 +26,7 @@ use crate::dates::{infer_temporal_granularity, parse_partial_date, parse_zoned};
 use crate::ratatui::print_ratatui_frame_to_stdout;
 use crate::scales::{Scale, ScaleType};
 use crate::select::SelectColumns;
-use crate::util;
+use crate::util::{self, ColorMode};
 use crate::{CliError, CliResult};
 
 const TYPICAL_COLS: usize = 35;
@@ -175,8 +175,11 @@ plot options:
                                [default: lin]
     --y-scale <scale>          Apply a scale to the y axis. Can be one of \"lin\" or \"log\".
                                [default: lin]
-    -C, --force-colors         Force colors even if output is not supposed to be able to
-                               handle them.
+    --color <when>             When to color the output using ANSI escape codes.
+                               Use `auto` for automatic detection, `never` to
+                               disable colors completely and `always` to force
+                               colors, even when the output could not handle them.
+                               [default: auto]
     -i, --ignore               Ignore values that cannot be correctly parsed.
 
 Common options:
@@ -216,7 +219,7 @@ struct Args {
     flag_y_max: Option<f64>,
     flag_x_scale: ScaleType,
     flag_y_scale: ScaleType,
-    flag_force_colors: bool,
+    flag_color: ColorMode,
     flag_ignore: bool,
 }
 
@@ -274,6 +277,7 @@ impl Args {
 
 pub fn run(argv: &[&str]) -> CliResult<()> {
     let args: Args = util::get_args(USAGE, argv)?;
+    args.flag_color.apply();
     let rconf = Config::new(&args.arg_input)
         .delimiter(args.flag_delimiter)
         .no_headers(args.flag_no_headers);
@@ -309,10 +313,6 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
 
     if matches!(args.flag_y_ticks, Some(n) if n.get() < 2) {
         Err("--y-ticks must be > 1!")?;
-    }
-
-    if args.flag_force_colors {
-        colored::control::set_override(true);
     }
 
     // Collecting data